refactor(fs): extract inline test module into dedicated file
Move the inline `#[cfg(test)] mod tests { ... }` block (313 lines)
from `ensure_file.rs` into `ensure_file/tests.rs` following
CODE_STYLE_GUIDE.md. No behavioral change.

Part of #358.
